src/coursera/week1/SortMaxPairwiseAlgorithm.java
package coursera.week1;

import java.io.*;
import java.util.*;
import java.math.BigInteger;

public class SortMaxPairwiseAlgorithm implements AlgorithmStrategy{

	public static void main(String[] args) {
		FastScanner scanner = new FastScanner(System.in);
		int n = scanner.nextInt();
		int[] numbers = new int[n];
		for (int i = 0; i < n; i++) {
			numbers[i] = scanner.nextInt();
		}
		System.out.println(getMaxPairwiseProduct(numbers));
	}

	static BigInteger getMaxPairwiseProduct(int[] numbers) {
		Arrays.sort(numbers);
		BigInteger a = BigInteger.valueOf(numbers[numbers.length-2]);
		BigInteger b = BigInteger.valueOf(numbers[numbers.length-1]);
		return a.multiply(b);
	}


	static class FastScanner {
		BufferedReader br;
		StringTokenizer st;

		FastScanner(InputStream stream) {
			try {
				br = new BufferedReader(new InputStreamReader(stream));
			} catch (Exception e) {
				e.printStackTrace();
			}
		}

		String next() {
			while (st == null || !st.hasMoreTokens()) {
				try {
					st = new StringTokenizer(br.readLine());
				} catch (IOException e) {
					e.printStackTrace();
				}
			}
			return st.nextToken();
		}

		int nextInt() {
			return Integer.parseInt(next());
		}
	}
}
